[SC-20615] Refactor tests to make it easier to extend them
Split the utility methods into a separate trait so that they can be
 reused in other tests.

The existing tests should still work, no behavior should have been changed

Author: Pranav Anand <anandpranavv@gmail.com>

GitOrigin-RevId: e430beaf69ebfd870fd955a5b8bc1c4cec4c6bb9
